refactor(ui): rename files
